requirements.txt
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_884c889c-96af-444f-bd6d-daddb5e9a462ykj3l5n_/croots/recipe/certifi_1655968814730/work/certifi
charset-normalizer==2.1.0
cycler==0.11.0
fonttools==4.34.4
idna==3.3
kiwisolver==1.4.4
matplotlib==3.5.2
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
tensorboardX==2.5.1
torch==1.13.0.dev20220725
torchaudio==0.14.0.dev20220603
torchinfo==1.7.0
torchsummary==1.5.1
torchvision==0.14.0.dev20220725
typing_extensions==4.3.0
urllib3==1.26.11
